chore History feature complete with minor issue, please validate!
I have added a new database operation called getchoredeadline so that I can relate chores completed or not, along with their associated deadline. Also, no errors are present however while testing it to see if the colour highlights work, I come across a logical error where it makes all the chores in the history to become red even if one is green. Could you all check if this is the same case for u and what fixes need to me made for this?